Extern census 149 -> 117: retype 28 modgfx resource externs u8[] -> u32[] to match their definers, drop tricky's dangling enemySightRange decl, resolve focusedNpc/gTitleMenuLinkInterface struct-prefix decls, unify the FamilyTable typedef name
All 28 modgfx consumers reach the sibling resource blobs only through
(u8*)(int) casts (the one bare pass in 154 gains the same cast every
sibling already uses); every rebuilt .o is byte-identical. tricky.c never
read enemySightRange - the extern was dangling. objseq_api.h picks up
object_fwd.h; gameloop_api.h typedefs its LinkInterface forward decl.
DOL bit-identical, fuzzy unchanged.

+++ src/dlls/modgfx/100/100.c
@@ -20,7 +20,7 @@
 /* effect id spawned by this DLL's modgfx emitter (spawnEffect textureAssetId arg). */
 #define DLL64_EFFECT_ID 0x5e0
 
-extern u8 lbl_80312D18[];
+extern u32 lbl_80312D18[];
 
 #define DLL64_PRIMARY_COMMAND_XZ 0.75f
 #define DLL64_UNIT_SCALE         1.0f
