+++ src/dlls/modgfx/102/102.c
@@ -13,7 +13,7 @@
 /* effect id spawned by this DLL's modgfx emitter (spawnEffect textureAssetId arg). */
 #define DLL66_EFFECT_ID 0x155
 
-extern u8 lbl_803131A8[];
+extern u32 lbl_803131A8[];
 
 void dll_66_func03(int sourceObj, int variant, int posSource, u32 flags)
 {
