+++ src/dlls/modgfx/95/95.c
@@ -22,7 +22,7 @@
 /* effect id spawned by this DLL's modgfx emitter (spawnEffect textureAssetId arg). */
 #define DLL5F_EFFECT_ID 0x48
 
-extern u8 lbl_80312650[];
+extern u32 lbl_80312650[];
 
 void dll_5F_func03(int sourceObj, int variant, int posSource, u32 flags)
 {
